+++ core/testsuite/src/test/java/com/blazebit/persistence/testsuite/JsonAndXmlTest.java
@@ -79,8 +79,8 @@ public class JsonAndXmlTest extends AbstractCoreTest {
     }
 
     // NOTE: DB2 crashes when executing this test with the GROUP_CONCAT based implementation
-    // NOTE: EclipseLink cant' handle multiple subquery select items... Only one expression can be declared in a SELECT clause of a subquery
-    // NOTE: DataNucleus cant' handle multiple subquery select items... Number of result expressions in subquery should be 1
+    // NOTE: EclipseLink can't handle multiple subquery select items... Only one expression can be declared in a SELECT clause of a subquery
+    // NOTE: DataNucleus can't handle multiple subquery select items... Number of result expressions in subquery should be 1
     @Test
     @Category({ NoDatanucleus.class, NoEclipselink.class, NoDB2.class })
     public void testToStringJson() {
@@ -103,8 +103,8 @@ public class JsonAndXmlTest extends AbstractCoreTest {
     }
 
     // NOTE: DB2 crashes when executing this test with the GROUP_CONCAT based implementation
-    // NOTE: EclipseLink cant' handle multiple subquery select items... Only one expression can be declared in a SELECT clause of a subquery
-    // NOTE: DataNucleus cant' handle multiple subquery select items... Number of result expressions in subquery should be 1
+    // NOTE: EclipseLink can't handle multiple subquery select items... Only one expression can be declared in a SELECT clause of a subquery
+    // NOTE: DataNucleus can't handle multiple subquery select items... Number of result expressions in subquery should be 1
     @Test
     @Category({ NoDatanucleus.class, NoEclipselink.class, NoDB2.class })
     public void testToStringXml() {
@@ -165,8 +165,8 @@ public class JsonAndXmlTest extends AbstractCoreTest {
     }
 
     // NOTE: DB2 crashes when executing this test with the GROUP_CONCAT based implementation
-    // NOTE: EclipseLink cant' handle multiple subquery select items... Only one expression can be declared in a SELECT clause of a subquery
-    // NOTE: DataNucleus cant' handle multiple subquery select items... Number of result expressions in subquery should be 1
+    // NOTE: EclipseLink can't handle multiple subquery select items... Only one expression can be declared in a SELECT clause of a subquery
+    // NOTE: DataNucleus can't handle multiple subquery select items... Number of result expressions in subquery should be 1
     @Test
     @Category({ NoDatanucleus.class, NoEclipselink.class, NoDB2.class })
     public void testToMultiset() {
@@ -189,8 +189,8 @@ public class JsonAndXmlTest extends AbstractCoreTest {
     }
 
     // NOTE: DB2 crashes when executing this test with the GROUP_CONCAT based implementation
-    // NOTE: EclipseLink cant' handle multiple subquery select items... Only one expression can be declared in a SELECT clause of a subquery
-    // NOTE: DataNucleus cant' handle multiple subquery select items... Number of result expressions in subquery should be 1
+    // NOTE: EclipseLink can't handle multiple subquery select items... Only one expression can be declared in a SELECT clause of a subquery
+    // NOTE: DataNucleus can't handle multiple subquery select items... Number of result expressions in subquery should be 1
     @Test
     @Category({ NoDatanucleus.class, NoEclipselink.class, NoDB2.class })
     public void testNestedToMultiset() {