+++ entity-view/impl/src/main/java/com/blazebit/persistence/view/impl/objectbuilder/transformer/correlation/CorrelatedSingularBatchTupleListTransformerFactory.java
@@ -22,6 +22,7 @@ import com.blazebit.persistence.view.CorrelationProviderFactory;
 import com.blazebit.persistence.view.impl.EntityViewConfiguration;
 import com.blazebit.persistence.view.impl.metamodel.ManagedViewTypeImplementor;
 import com.blazebit.persistence.view.impl.objectbuilder.Limiter;
+import com.blazebit.persistence.view.impl.objectbuilder.transformer.NullListTupleTransformer;
 import com.blazebit.persistence.view.impl.objectbuilder.transformer.TupleListTransformer;
 
 import java.util.Map;
@@ -39,8 +40,11 @@ public class CorrelatedSingularBatchTupleListTransformerFactory extends Abstract
     }
 
     @Override
-    public TupleListTransformer create(ParameterHolder<?> parameterHolder, Map<String, Object> optionalParameters, EntityViewConfiguration entityViewConfiguration) {
-        return new CorrelatedSingularBatchTupleListTransformer(entityViewConfiguration.getExpressionFactory(), correlator, viewRootType, embeddingViewType, correlationResult, correlationProviderFactory, attributePath, fetches, correlatesThis, viewRootIndex, embeddingViewIndex, tupleIndex, batchSize, correlationBasisType, correlationBasisEntity, limiter, entityViewConfiguration);
+    public TupleListTransformer create(ParameterHolder<?> parameterHolder, Map<String, Object> optionalParameters, EntityViewConfiguration config) {
+        if (!config.hasSubFetches(attributePath)) {
+            return new NullListTupleTransformer(tupleIndex, correlator.getElementOffset());
+        }
+        return new CorrelatedSingularBatchTupleListTransformer(config.getExpressionFactory(), correlator, viewRootType, embeddingViewType, correlationResult, correlationProviderFactory, attributePath, fetches, correlatesThis, viewRootIndex, embeddingViewIndex, tupleIndex, batchSize, correlationBasisType, correlationBasisEntity, limiter, config);
     }
 
 }