+++ entity-view/testsuite/src/test/java/com/blazebit/persistence/view/testsuite/multiset/MultisetFetchCollectionsBenchmarkTest.java
@@ -156,8 +156,8 @@ public class MultisetFetchCollectionsBenchmarkTest extends AbstractEntityViewTes
     }
 
     // NOTE: DB2 crashes when executing this test with the GROUP_CONCAT based implementation
-    // NOTE: EclipseLink cant' handle multiple subquery select items... Only one expression can be declared in a SELECT clause of a subquery
-    // NOTE: DataNucleus cant' handle multiple subquery select items... Number of result expressions in subquery should be 1
+    // NOTE: EclipseLink can't handle multiple subquery select items... Only one expression can be declared in a SELECT clause of a subquery
+    // NOTE: DataNucleus can't handle multiple subquery select items... Number of result expressions in subquery should be 1
     @Test
     @Category({ NoDB2.class, NoDatanucleus.class, NoEclipselink.class })
     @BenchmarkOptions(benchmarkRounds = 5, warmupRounds = 4)
@@ -166,8 +166,8 @@ public class MultisetFetchCollectionsBenchmarkTest extends AbstractEntityViewTes
     }
 
     // NOTE: DB2 crashes when executing this test with the GROUP_CONCAT based implementation
-    // NOTE: EclipseLink cant' handle multiple subquery select items... Only one expression can be declared in a SELECT clause of a subquery
-    // NOTE: DataNucleus cant' handle multiple subquery select items... Number of result expressions in subquery should be 1
+    // NOTE: EclipseLink can't handle multiple subquery select items... Only one expression can be declared in a SELECT clause of a subquery
+    // NOTE: DataNucleus can't handle multiple subquery select items... Number of result expressions in subquery should be 1
     @Test
     @Category({ NoDB2.class, NoDatanucleus.class, NoEclipselink.class })
     @BenchmarkOptions(benchmarkRounds = 5, warmupRounds = 4)