+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Changes that happened in releases
 * Treat entity view mapping parameters always as singular in annotation processor
 * Import and properly render parameterized type arguments in entity view annotation processor
 * Fix entity view by id remove error on cascade delete of one-to-one mapped by association
+* Fix dynamic fetch support for `MULTISET` fetched mappings
 
 ### Backwards-incompatible changes